Cleaned up code and moved the arrows to their own component
